move blizzard service to seperate source file
